+++ shared/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.37] - 2023-03-01
+
+- add pokt --> libp2p crypto helpers
+
 ## [0.0.0.36] - 2023-02-28
 
 - Move `StakeStatus` into `actor.proto`